Remove getter method.
In general, getters should not be used in Python. You should just use the
member variable directly, as I've updated this to do. In the case that
computation needs to be done on the fly, the property decorator should be
used.

Also, I've introduced temporary variables in the starburst initialization
to improve readability. I'm doing this in the same commit largely because
splitting a line on square brackets kills me.